(refactoring) Move TSchemeCacheHelpers to core/tx/scheme_cache KIKIMR-20673
