+++ ydb/core/tx/datashard/change_exchange_split.cpp
@@ -6,6 +6,8 @@
 #include <ydb/core/base/tablet_pipe.h>
 #include <ydb/core/persqueue/events/global.h>
 #include <ydb/core/persqueue/writer/source_id_encoding.h>
+#include <ydb/core/tx/scheme_cache/helpers.h>
+#include <ydb/core/tx/scheme_cache/scheme_cache.h>
 #include <ydb/public/lib/base/msgbus_status.h>
 
 #include <ydb/library/actors/core/actor_bootstrapped.h>
@@ -155,7 +157,10 @@ private:
 
 }; // TCdcPartitionWorker
 
-class TCdcWorker: public TActorBootstrapped<TCdcWorker>, private TSchemeCacheHelpers {
+class TCdcWorker
+    : public TActorBootstrapped<TCdcWorker>
+    , private NSchemeCache::TSchemeCacheHelpers
+{
     TStringBuf GetLogPrefix() const {
         if (!LogPrefix) {
             LogPrefix = TStringBuilder()
